[core] Public RESTTokenFileIO.validToken which can be invoked by native engines to get the token (#5063)
